requirements.txt
streamlit==1.31.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.2.0
chromadb==0.4.22
sentence-transformers==2.7.0
torch==2.1.2
numpy==1.24.3
requests==2.31.0
google-generativeai==0.3.2
markdown==3.5.1
gTTS==2.5.0
beautifulsoup4==4.12.3
youtube-transcript-api==0.6.2
langdetect==1.0.9
pytesseract==0.3.10